+++ fj-daogen-base/src/main/java/org/fugerit/java/daogen/base/config/DaogenCatalogConstants.java
@@ -44,6 +44,9 @@ public class DaogenCatalogConstants {
 	public static final String GEN_PROP_PACKAGE_SPRING_REST_LOAD = 	"package-spring-rest-load";
 	public static final String GEN_PROP_PACKAGE_REST_LOAD = 	"package-rest-load";
 	public static final String GEN_PROP_BASE_REST_SERVICE = 	"base-rest-service";
+
+	public static final String GEN_PROP_PACKAGE_JUNIT4_MODEL = 	"package-junit4-model";
+
 	public static final String GEN_PROP_STRUCT_PREFIX = 	"struct-prefix";
 	public static final String GEN_PROP_STRUCT_PREFIX_DEFAULT = 	"OBJ_";
 	
@@ -84,6 +87,8 @@ public class DaogenCatalogConstants {
 	public static final String PREFIX_FACADE_DEF = "Entity";
 	public static final String SUFFIX_FACADE_DEF = "Facade";
 	public static final String PREFIX_FACADE_IMPL_DATA = "Data";
+
+	public static final String SUFFIX_JUNIT4_MODEL = "Junit4ModelTest";
 	
 	public static String modelName( DaogenCatalogEntity entity ) {
 		return PREFIX_MODEL+entity.toClassName();
@@ -111,6 +116,10 @@ public class DaogenCatalogConstants {
 	
 	public static String finderlName( DaogenCatalogEntity entity ) {
 		return entity.toClassName()+SUFFIX_FINDER;
+	}
+
+	public static String junit4ModelName( DaogenCatalogEntity entity ) {
+		return entity.toClassName()+SUFFIX_JUNIT4_MODEL;
 	}
 	
 	public static String facadeDefName( DaogenCatalogEntity entity ) {